Release 0.1 (m)

+++ sources/read_experience_data.cpp
@@ -5,7 +5,7 @@ void MainWindow::read_experience_data(QFile &romfile)
     QDataStream read(&romfile);
     romfile.seek(0x7DE453);
 
-    for (short i=0;i<6;i++){
+    for(uint8_t i=0;i<6;i++){
         read>>experience_param_1[i];
         if(i<4){
             read>>experience_param_2[i];
